--- src/universal/api/index.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-import { URL } from '../consts/apiConsts';
-
-export default function init() {
-  axios.defaults.baseURL = URL;
-}
